Removed old code and all references to it, switched to new API
