+++ allennlp/semparse/domain_languages/__init__.py
@@ -1,4 +1,5 @@
 from allennlp.semparse.domain_languages.domain_language import (DomainLanguage, ParsingError,
-                                                                ExecutionError, predicate)
+                                                                ExecutionError, predicate,
+                                                                START_SYMBOL)
 from allennlp.semparse.domain_languages.nlvr_language import NlvrLanguage
 :rom allennlp.semparse.domain_languages.quarel_language import QuaRelLanguage
