+++ allennlp/semparse/worlds/__init__.py
@@ -1,4 +1,3 @@
-from allennlp.semparse.worlds.nlvr_world import NlvrWorld
 from allennlp.semparse.worlds.wikitables_world import WikiTablesWorld
 from allennlp.semparse.worlds.wikitables_variable_free_world import WikiTablesVariableFreeWorld
 from allennlp.semparse.worlds.atis_world import AtisWorld
